docs(python): improve error handling section
The error handling section does not currently show where the `ApiException`
class comes from, so it's difficult to determine how to define it. Furthermore,
the `ApiException` class does not actually have an `info` field, so the docs
are misleading.

This commit adds an import statement for the class to make it clear and removes
mentions of the non-existent `info` field.