docs: update Java deployment and publishing (#69)
This commit updates the documentation of the Java deployment and
publishing process, since it has changed quite much recently.
The main update is the migration from the OSSRH service (which has
been sunset) to the new Central Publishing process.
The secondary change is related to the slow transition from Travis
to GitHub Actions. We've started to prefer the latter especially for
new projects, so it's time to adjust the docs too. It still contains the
Travis related steps as it's a valid option for CI/CD, it's just not the one
that we strongly recommend anymore.

Signed-off-by: Norbert Biczo <pyrooka@users.noreply.github.com>